完成userInfoController并测试

+++ api/src/main/java/com/yangweiye/api/UserInfoService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface UserInfoService {
     List<UserInfo> findAll();
 
-    UserInfo findById(Integer id);
+    UserInfo findById(Long id);
 
     Integer insertUser(UserInfo user);
